# frontend/openapi_spec_pretty.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8086",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/agents": {
      "post": {
        "tags": [
          "Agents"
        ],
        "summary": "Submit a request to an agent",
        "operationId": "submitRequest",
        "parameters": [
          {
            "name": "domainLanguage",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Domain language hint used to drive localisation of responses.",
              "enum": [
                "fr",
                "en"
              ],
              "example": "fr"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {}
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Request submitted successfully",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {}
          },
          "500": {
            "description": "Internal server error",
            "content": {}
          }
        }
      }
    },
...
    "/agents/{issueId}": {
      "get": {
        "tags": [
          "Agents"
        ],
        "summary": "Get details of an agent-created issue",
        "operationId": "getIssue",
        "parameters": [
          {
            "name": "issueId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "domainLanguage",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Domain language hint used to drive localisation of responses.",
              "enum": [
                "fr",
                "en"
              ],
              "example": "fr"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AgentIssueDto"
                }
              }
            }
          }
        }
      }
    },
    "/agents/templates": {
      "get": {
        "tags": [
          "Agents"
        ],
        "summary": "Get available agent templates",
        "operationId": "listTemplates",
        "parameters": [
          {
            "name": "domainLanguage",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Domain language hint used to drive localisation of responses.",
              "enum": [
                "fr",
                "en"
              ],
              "example": "fr"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AgentTemplateDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/agents/mailto": {
      "get": {
        "tags": [
          "Agents"
        ],
        "summary": "Get mailto link for fallback",
        "operationId": "getMailto",
        "parameters": [
          {
            "name": "agentId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "domainLanguage",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Domain language hint used to drive localisation of responses.",
              "enum": [
                "fr",
                "en"
              ],
              "example": "fr"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/agents/activity": {
      "get": {
        "tags": [
          "Agents"
        ],
        "summary": "Get recent agent activity",
        "operationId": "listActivity",
        "parameters": [
          {
            "name": "domainLanguage",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Domain language hint used to drive localisation of responses.",
              "enum": [
                "fr",
                "en"
              ],
              "example": "fr"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AgentActivityDto"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AgentIssueDto": {
        "type": "object",
        "description": "Detailed view of an agent-created issue",
        "properties": {
          "issueId": {
            "type": "string",
            "description": "Issue ID"
          },
          "number": {
            "type": "integer",
            "format": "int32",
            "description": "Issue Number"
          },
          "title": {
            "type": "string",
            "description": "Issue Title"
          },
          "url": {
            "type": "string",
            "description": "Issue URL"
          },
          "status": {
            "type": "string",
            "description": "Issue Status (open/closed)"
          },
          "labels": {
            "type": "array",
            "description": "Issue Labels",
            "items": {
              "type": "string"
            }
          },
          "workflowState": {
            "type": "string",
            "description": "Workflow State"
          },
          "previewUrl": {
            "type": "string",
            "description": "Preview URL"
          },
          "promptVisibility": {
            "type": "string",
            "description": "Prompt Visibility",
            "enum": [
              "PUBLIC",
              "PRIVATE"
            ]
          },
          "promptSummary": {
            "type": "string",
            "description": "Prompt Summary (if public)"
          },
          "comments": {
            "type": "array",
            "description": "Discussion comments",
            "items": {
              "$ref": "#/components/schemas/IssueCommentDto"
            }
          }
        }
      },
      "IssueCommentDto": {
        "type": "object",
        "description": "Issue comment",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Comment identifier"
          },
          "author": {
            "type": "string",
            "description": "Comment author"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          },
          "body": {
            "type": "string",
            "description": "Markdown content"
          }
        }
      },
      "AgentAttributeDto": {
        "type": "object",
        "description": "Custom attribute definition",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "options": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "AgentTemplateDto": {
        "type": "object",
        "description": "Configuration template for an agent",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the agent"
          },
          "name": {
            "type": "string",
            "description": "Localized name of the agent"
          },
          "description": {
            "type": "string",
            "description": "Localized description"
          },
          "icon": {
            "type": "string",
            "description": "Icon identifier (MDI)"
          },
          "promptTemplates": {
            "type": "array",
            "description": "Available prompt templates",
            "items": {
              "$ref": "#/components/schemas/PromptTemplateDto"
            }
          },
          "tags": {
            "type": "array",
            "description": "Associated tags",
            "items": {
              "type": "string"
            }
          },
          "allowedRoles": {
            "type": "array",
            "description": "Allowed roles to access this agent",
            "items": {
              "type": "string"
            }
          },
          "publicPromptHistory": {
            "type": "boolean",
            "description": "Whether the prompt history is public"
          },
          "allowTemplateEditing": {
            "type": "boolean",
            "description": "Whether user can edit the selected prompt template"
          },
          "mailTemplate": {
            "$ref": "#/components/schemas/MailTemplateDto",
            "description": "Mailto template configuration"
          },
          "attributes": {
            "type": "array",
            "description": "Custom attributes for the agent",
            "items": {
              "$ref": "#/components/schemas/AgentAttributeDto"
            }
          }
        }
      },
      "MailTemplateDto": {
        "type": "object",
        "description": "Mailto configuration",
        "properties": {
          "to": {
            "type": "string"
          },
          "subject": {
            "type": "string"
          },
          "body": {
            "type": "string"
          }
        }
      },
      "PromptTemplateDto": {
        "type": "object",
        "description": "Prompt template option",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the prompt template"
          },
          "title": {
            "type": "string",
            "description": "Localized title of the template"
          },
          "content": {
            "type": "string",
            "description": "Prompt content in markdown/inline format"
          }
        }
      },
      "AgentActivityDto": {
        "type": "object",
        "description": "Summary of a recent agent request",
        "properties": {
          "id": {
            "type": "string",
            "description": "Issue ID"
          },
          "type": {
            "type": "string",
            "description": "Request Type",
            "enum": [
              "FEATURE",
              "QUESTION"
            ]
          },
          "issueUrl": {
            "type": "string",
            "description": "Issue URL"
          },
          "status": {
            "type": "string",
            "description": "Issue Status"
          },
          "promptVisibility": {
            "type": "string",
            "description": "Prompt Visibility",
            "enum": [
              "PUBLIC",
              "PRIVATE"
            ]
          },
          "promptSummary": {
            "type": "string",
            "description": "Prompt Summary (if public)"
          },
          "commentsCount": {
            "type": "integer",
            "format": "int32",
            "description": "Number of discussion comments"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}